HBAPP-94 Fix size of prompt shadow blur. Fix theme between server/client. Move tab files in folder. Rename query charm menu.
