+++ app/page.tsx
@@ -1,9 +1,8 @@
-import { nanoid } from '@/lib/utils'
 import { ChatContainer } from '@/components/chat-container'
 
 export const runtime = 'edge'
 
 export default function IndexPage() {
-  const id = nanoid()
-  return <ChatContainer id={id} />
+  // @TODO Change to AppContainer
+  return <ChatContainer />
 }
