+++ components/features/menus/bots/menu-create-bot.tsx
@@ -10,13 +10,13 @@ import { Separator } from '@/components/ui/separator'
 import { Button } from '@/components/ui/button'
 import { Tabs } from '@/components/ui/tabs'
 import { I_Attention_State, I_Knowledge_State, I_LLM_Init_Options, I_ModelConfigs, I_Model_State, I_PromptTemplates, I_Prompt_State, I_RAGPromptTemplates, I_Response_State, I_ServiceApis, I_System_State, I_Text_Settings, T_InstalledTextModel, useHomebrew } from '@/lib/homebrew'
-import { AttentionTab, defaultState as defaultAttentionState } from '@/components/features/menus/bots/tab-attention'
-import { PerformanceTab, defaultState as defaultPerformanceState } from '@/components/features/menus/bots/tab-performance'
-import { ModelTab, defaultState as defaultModelState } from '@/components/features/menus/bots/tab-model'
-import { SystemTab, defaultState as defaultSystemState } from '@/components/features/menus/bots/tab-system'
-import { PromptTab, defaultState as defaultPromptState } from '@/components/features/menus/bots/tab-prompt'
-import { KnowledgeTab, defaultState as defaultKnowledgeState } from '@/components/features/menus/bots/tab-knowledge'
-import { ResponseTab, defaultState as defaultResponse } from '@/components/features/menus/bots/tab-response'
+import { AttentionTab, defaultState as defaultAttentionState } from '@/components/features/menus/tabs/tab-attention'
+import { PerformanceTab, defaultState as defaultPerformanceState } from '@/components/features/menus/tabs/tab-performance'
+import { ModelTab, defaultState as defaultModelState } from '@/components/features/menus/tabs/tab-model'
+import { SystemTab, defaultState as defaultSystemState } from '@/components/features/menus/tabs/tab-system'
+import { PromptTab, defaultState as defaultPromptState } from '@/components/features/menus/tabs/tab-prompt'
+import { KnowledgeTab, defaultState as defaultKnowledgeState } from '@/components/features/menus/tabs/tab-knowledge'
+import { ResponseTab, defaultState as defaultResponse } from '@/components/features/menus/tabs/tab-response'
 import { useMemoryActions } from '@/components/features/crud/actions'
 import { toast } from 'react-hot-toast'
