+++ lib/utils.ts
@@ -45,6 +45,6 @@ export function formatDate(input: string | number | Date): string {
 
 export const constructMainBgStyle = (currentTheme: string | undefined) => {
   const bgColor = `${currentTheme === 'light' ? 'bg-neutral-200' : 'bg-neutral-900'}`
-  const wrapperStyle = `flex flex-1 flex-col items-center self-stretch ${bgColor}`
+  const wrapperStyle = `flex w-full flex-1 flex-col items-center self-stretch ${bgColor}`
   return wrapperStyle
 }
